Change buffer to non-copy-constructible
I don't fully understand the TMP stuff around the zipped_provider_t. But
it seems evident that I can't make the zip work without it holding a copy
of its arguments. So if the buffer is to be a no-copy object, then you
can't include a unique array in a zipped provider. That's OK I think.